version 0.1-1

+++ src/CToR.c
@@ -19,7 +19,7 @@ SEXP c_to_r_boolean(int c_boolean){
 SEXP c_to_r_integer(long c_long){
     SEXP r_val;
     PROTECT(r_val = allocVector(INTSXP, 1));
-    INTEGER(r_val)[0] = c_long;
+    INTEGER(r_val)[0] = (int)c_long;
     UNPROTECT(1);
     return r_val;
 }
